Fix Pyright errors: missing imports, type annotations, and guards
- dispatcher: add missing save_pending_start/discard_pending_start imports
- fuzzy_time: fix _hour_in_range parameter type (int -> float)
- nap/sleep: add assert guards after parse_time_range for type narrowing
- terminal_ui: fix choose_from_list parameter name to match base class
- calendar_events: add None guards for get_events() return value
- test_events: add assertIsNotNone before assertIn calls

+++ dailydriver/cli/search/fuzzy_time.py
@@ -20,7 +20,7 @@ TIME_RANGES = {
     "night": (16, 4),   # 16:00 to 04:00 next day
 }
 
-def _hour_in_range(hour: int, start: int, end: int) -> float:
+def _hour_in_range(hour: float, start: int, end: int) -> float:
     """Return distance in hours from hour to the range, 0 if inside."""
     if start <= end:
         if start <= hour < end:
